requirements.txt

fastapi
uvicorn
requests
httpx
jinja2
pillow
sqlalchemy
types-requests
python-multipart
pyjwt
